Minor refactor to AttackComponent
Component now has some editor customizable storage for attacks. There was a bug where the attack data wouldn't load by the time it's needed. Now we just serialize all the attacks we'll need and reference them as needed, based on the entity's state. This may need changing in the future but it's good for now.