Player no longer starts the game in jump land
Bug occurred due to physics collisions not being queried at the first frame of gameplay. Game now starts the player in an init state. The init state immediately transitions the player to the idle state, at which point physics collisions are properly detected and the player doesn't immediately go into the jump and jump land states.